requirements.txt
undetected-chromedriver
selenium
python-dotenv
fake_useragent
blinker==1.6.3     # selenium-wire/mitmproxy needs _saferef

# Add your project dependencies here
